Moved danceCritiqueForm to its own component
