+++ App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Provider } from 'react-redux'
 import { createStore } from 'redux'
 
-import DanceCritiqueForm from './danceCritiqueForm'
+import DanceCritiqueForm from './components/DanceCritiqueForm'
 import rootReducer from './rootReducer.js'
 
 const store = createStore(rootReducer)
